renaming ditcustomerpage
